requirements.txt
# Neo4j Demo Toolkit - Python Dependencies

# Neo4j Python Driver
neo4j>=5.14.0

# Environment variable management
python-dotenv>=1.0.0
